+++ muon/film.h
@@ -17,7 +17,8 @@ class Film {
  public:
   // Initializes a new Film with a given width and height and a path to an
   // output file, which should be a png.
-  Film(int width, int height, int samples_per_pixel, std::string output_file)
+  Film(size_t width, size_t height, size_t samples_per_pixel,
+       std::string output_file)
       : width_(width),
         height_(height),
         samples_per_pixel_(samples_per_pixel),
@@ -28,15 +29,15 @@ class Film {
   Film &operator=(Film &&other) = default;
 
   // Sets a specific pixel coordinate to a given color.
-  void SetPixel(int x, int y, glm::vec3 color);
+  void SetPixel(size_t x, size_t y, glm::vec3 color);
 
   // Writes the sampled output to disk.
   void WriteOutput();
 
  private:
-  int width_;
-  int height_;
-  int samples_per_pixel_;
+  size_t width_;
+  size_t height_;
+  size_t samples_per_pixel_;
   std::string output_file_;
   std::vector<std::vector<glm::vec3>> accumulator_;
   cimg_library::CImg<unsigned char> output_;
